MRG: report checksum file download failures (#92)
- Fixes #70 

This PR reports any failures brought about by issues downloading or
parsing the md5sum file within `gbsketch`. These cases were not
previously included in the failures file. This is the issue in #70.

Now that we have a separate output file, I thought it might be useful to
also pull out cases where the md5sum does not match expectation, and
write those to the md5sum failure file as well (added in
6755296f80d0a859ecc1a81f73e94418b7bfd120). These were previously just
written to the download failures file, with no explanation for _why_ the
download failed (e.g. whether it was a download issue or an md5sum
matching issue).

There is no straightforward way to test the checksum failures file via
`gbsketch`, but we use the mismatch md5sum case to test it via
`urlsketch`.

There is a "major" change this PR brings: the `--checksum-fail` file is
now required for `gbsketch`. I think it's not possible to make this
optional with the current strategy, because otherwise the errors for
failed md5sum file download/parsing would not be properly captured
anywhere. I have made `--checksum-fail` optional for `urlsketch`,
because there folks need to manually provide expected md5sums, so we
don't have md5sum file download issues and often we won't even have
md5sums to check.