MRG: add generic support for signature building (#101)
This PR introduces `BuildCollection`, `BuildRecord`, `BuildManifest`
that can be used to build sigs. Since `BuildCollection` is restricted to
a set of template sigs that get modified with data, I also introduce
`MultiBuildCollection` as a simple wrapper to hold `BuildCollections`.

These mirror `Collection`, `Record`, `Manifest` in sourmash core, with
the following major differences:
1. `BuildRecord` requires less information than `Record`, allowing build
collection to be template signatures that are filled while reading
files. However, `BuildRecord` contains placeholders for all `Record`
items, which means that we can update these while building sigs and
write a standard manifest via `BuildManifest`. This reduces the work we
need to do when writing sigs, as it is no longer necessary to build a
new `Record` from each signature in order to build the `Manifest` to
write.
3. `BuildCollection` contains methods for `build_sigs_from_data`,
`build_sigs_from_file`, and `build_singleton_sigs` which fill template
sigs from data, file, or fastx sequence record, respectively. These take
care of updating all `BuildRecord` info, sig name, and sig filename, and
so should make sketching applications much simpler to write.
4. `BuildCollection` can be filtered by `Params`. This is for future
use, so we can read in a set of existing signatures and filter out any
template sigs that do not need to be built.
5. `BuildCollection` and `BuildManifest`contain `async` writer functions
to write the signatures/ manifest to a zipfile. I would imagine these
stay here in directsketch, and we modify them to be `sync` for use in
`manysketch`. I do think it useful to implement these within the struct
because, for example, we need to make sure the datetime is set properly,
make sure to update a record's `internal_location` to reflect the file
location (which is currently influenced by md5sum occurrences), etc.
6. `MultiBuildCollection` is just a simple Vec<BuildCollection>.
However, it made it a bit easier to send all sigs associated with a
single accession to the writer channel while restricting a single
`BuildCollection` to a set of template sigs that get modified.

To Do:
- [x]  method to populate the remaining `BuildRecord` fields.
- [x] check that `send_sigs` is set up to write from BuildCollection
similarly / make sure sigs are written regularly. Previously, the
`send_sigs` channel size of 4 meant 4 sigs. Here now it's 4
`BuildCollection`s, which means 4 accessions/files or 2 accessions with
both protein, dna.
- [x] update sig, manifest writing to better use `BuildCollection`
- [x] **protein sig writing is currently not correct -- md5sum is
different than expected. Fix!**
- this was `k=k*3` rearing it's ugly head again. Need to use it ONLY for
sigs, not for manifest.
- [x] remove changes that will be used for zip batching (separate PR)