MRG: remove `BuildParams`, filter via manifest / `Select` approaches (#127)
This PR replaces the `BuildParams` manual hashing + filtering with
manifest filtering and selection (e.g. of moltypes) via a `Select`
framework. This is more in tune with the approaches in sourmash core. I
did add `MultiSelect` here, since we want to keep all templates that
match sets of selection parameters.

Internal improvements to the framework introduced in 0.4.0:
- All of the parameter parsing + checks from
`BuildParams`/`BuildParamsSet` is now in
`BuildRecord`/`BuildCollection`. We now parse the parameter string
directly into a `BuildCollection`, rather than going through
`BuildParams` as an intermediary.
- To manage finding and filtering existing signatures, we now select on
the `BuildCollection` directly via `MultiSelect`, e.g. for moltype
filtering. We can also filter a manifest with another manifest by using
PartialEq for `BuildRecord`s. This replaces the prior approach of
keeping a `BuildParamsSet` and hashing `BuildParams` manually for
comparisons.
- We continue to use sourmash core `ComputeParameters` to actually build
template signatures.
- Instead of handling DNA, protein as separate collections, manage them
jointly by allowing selection on moltype and adding to DNA sigs when we
have DNA, prot sigs when we have proteins. This reduces complexity of
use and has the added benefit of easier addition of translated sigs if
we want to add that in the future.
- Fixes #113