MRG: `gbsketch` streaming processing (#230)
This PR enables processing directly on the stream (`gbsketch` only
here), meaning we do not have to download the entire file prior to
processing! This should reduce memory requirements, since we no longer
need to hold full genomes in memory.

Other internal changes:
- We now read gbsketch input directly to an `AccessionData` struct,
rather than having a separate bespoke struct for `gbsketch`. We also use
`AccessionData` directly instead of building a `FailedDownload`. These
changes simplify processing and will better enable code reuse between
`gbsketch` and `urlsketch`.
- After finding the fetch urls via the dehydrated file, we then add them
to the `AccessionData` structs and use `stream_and_process_with_retry`
to stream and save/sketch the data. This function should be reusable for
`urlsketch`, if we can add md5sum checks. Not really sure how useful
having that field is for `urlsketch` anyway, since folks need to go find
the md5sum for their files.
- added csv writing trait, which allowed for consolidation of csv
writing handles
- enabled option to write `urlsketch` file from gbsketch, meaning we can
skip finding the links a second time. This will probably be more useful
after I get `urlsketch` updated with streaming methods to prevent memory
issues.
- use `.incomplete` for incomplete fastas, rename once finished. Current
approach deletes the incomplete fastas when we're going to retry or if
fastx parsing failed, but not if `Ctrl-C` is encountered. Happy to
modify if folks want that behavior to change. We truncate a file on
opening, so there's no risk of appending to existing *incomplete files
on subsequence runs.

Fixes:
- #228 
- #165 
- #47