MRG: optionally allow empty sig zips; write list of batch outputs to file (#249)
This PR fixes the two issues described in #248.

- First, we add a new option, `--allow-completed`. This will allow
snakemake to think that abatched workflow succeeded, even if no new
signatures were written **as long as we find some existing signatures in
the batches**. This will likely be most useful when restarting batches,
as if there are only _failing_ entries left, there will be no signatures
to write. Note that you will then want to go back and look at the
failures csv file to make sure those are truly failures (and if not, to
retry building those separately).
 
- Second, we add a new output for batching, the batchlist file. This
file will be "{output}.batchlist.txt", so if you're writing sigs to
`simple.zip`, this will be `simple.zip.batchlist.txt`. This will contain
all batches detected, whether they were previously written or added
during the current run. This file can be the output given to the
directsketch snakemake rule, since it will be created for all batched
runs. It should then be able to be used with `sourmash sig cat` or
`sourmash sig collect` to combine the zipfiles into a single zip or
collect them into a standalone manifest.

- fixes #248